% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_pmf_percentage_profiles.R
\name{adjust_pmf_percentage_profiles}
\alias{adjust_pmf_percentage_profiles}
\title{Function to adjust the \code{percentage_of_factor_total} entries from the
\code{\link{tidy_pmf_profiles}} output to ensure the factor sums equal 100 \%.}
\usage{
adjust_pmf_percentage_profiles(df)
}
\arguments{
\item{df}{Data frame/tibble from \code{\link{tidy_pmf_profiles}}.}
}
\value{
Tibble.
}
\description{
Function to adjust the \code{percentage_of_factor_total} entries from the
\code{\link{tidy_pmf_profiles}} output to ensure the factor sums equal 100 \%.
}
\author{
Stuart K. Grange
}
