% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_missing_mass.R
\name{calculate_missing_mass}
\alias{calculate_missing_mass}
\title{Function to calculate particulate matter missing mass.}
\usage{
calculate_missing_mass(df, digits = 6)
}
\arguments{
\item{df}{Input data frame.}

\item{digits}{Number of digits for logic test.}
}
\value{
Tibble.
}
\description{
\code{\link{calculate_missing_mass}} will add one more observation/row to
the input data frame/tibble.
}
\examples{

# Example input
data_example <- tibble::tribble(
 ~variable,        ~value,
 "mass",           39.7,  
 "ec",             6.1,   
 "om",             9.2,   
 "nitrage",        3.3,   
 "ammonium",       1.4,   
 "sulphate",       3.3,   
 "mineral_dust",   2.7,   
 "trace_elements", 6.8
 )
 
# Determine missing mass
calculate_missing_mass(data_example)

}
\seealso{
\code{\link{calculate_contributions}}
}
\author{
Stuart K. Grange
}
