% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pmf_input_data.R
\name{export_pmf_input_data}
\alias{export_pmf_input_data}
\title{Function to export a data table for use with the EPA PMF tool.}
\usage{
export_pmf_input_data(df, file)
}
\arguments{
\item{df}{Data frame to export.}

\item{file}{File name to export data to.}
}
\value{
Invisible tibble.
}
\description{
Function to export a data table for use with the EPA PMF tool.
}
\author{
Stuart K. Grange
}
