% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmfprof.R
\name{formatprof}
\alias{formatprof}
\title{formatprof}
\usage{
formatprof(base, cn = NULL, constr = NULL, scale = F)
}
\arguments{
\item{base}{base profile data read from *_profiles.txt}

\item{cn}{optional column names}

\item{constr}{constrained output}

\item{scale}{whether to row standardize (default = F)}
}
\description{
\code{formatprof} Formats PMF profile output
}
