% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_re_exports.R
\name{ggplot2 functions}
\alias{ggplot2 functions}
\title{Pseudo-function to re-export \strong{ggplot2}'s common functions.}
\description{
Pseudo-function to re-export \strong{ggplot2}'s common functions.
}
