% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_epa_pmf_outputs.R
\name{list_epa_pmf_outputs}
\alias{list_epa_pmf_outputs}
\title{Function to list files/paths of the outputs created by the EPA PMF tool.}
\usage{
list_epa_pmf_outputs(directory)
}
\arguments{
\item{directory}{Directory where EPA PMF files are located.}
}
\value{
Named list containing \strong{fs} paths.
}
\description{
Function to list files/paths of the outputs created by the EPA PMF tool.
}
\author{
Stuart K. Grange
}
