% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_plotting_functions.R
\name{plot_pmf_mass_factor_contributions}
\alias{plot_pmf_mass_factor_contributions}
\title{Function to plot PMF mass factor contributions using data from 
\code{\link{extract_pmf_mass_factor_contributions}}.}
\usage{
plot_pmf_mass_factor_contributions(df, round = 1)
}
\arguments{
\item{df}{Tibble from \code{\link{extract_pmf_mass_factor_contributions}}.}

\item{round}{Number of digits used for percentage labels.}
}
\value{
\strong{ggplot2} with bar geometries.
}
\description{
Function to plot PMF mass factor contributions using data from 
\code{\link{extract_pmf_mass_factor_contributions}}.
}
\author{
Stuart K. Grange
}
