% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_plotting_functions.R
\name{plot_pmf_q_robust}
\alias{plot_pmf_q_robust}
\title{Function to plot PMF Q robust scores from 
\code{\link{read_pmf_diagnostics}}'s \code{base_run_summary_table} table.}
\usage{
plot_pmf_q_robust(df)
}
\arguments{
\item{df}{Tibble from \code{\link{read_pmf_diagnostics}}.}
}
\value{
ggplot2 with point geometries.
}
\description{
Function to plot PMF Q robust scores from 
\code{\link{read_pmf_diagnostics}}'s \code{base_run_summary_table} table.
}
\author{
Stuart K. Grange
}
