% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmfcont.R
\name{plotconstrain}
\alias{plotconstrain}
\title{plotconstrain}
\usage{
plotconstrain(contrib, scale = NULL)
}
\arguments{
\item{contrib}{pmf contributions}

\item{scale}{whether to scale}
}
\description{
plotconstrain
}
