% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_species_contributions_to_factor.R
\name{pmf_species_contributions_to_factor}
\alias{pmf_species_contributions_to_factor}
\title{Function to calculate the species contribution to the factor.}
\usage{
pmf_species_contributions_to_factor(df, species_mass = "mass")
}
\arguments{
\item{df}{Data frame/tibble from \code{\link{tidy_pmf_profiles}}.}

\item{species_mass}{Name of the species which was used as the "total 
variable".}
}
\value{
Tibble.
}
\description{
This function is useful for when the contribution/percentage of an input
species is desired. \code{species_contributions_to_factor} will exclude the
"total variable" species (usually PM mass) from its calculation.
}
\author{
Stuart K. Grange
}
