% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmfcont.R
\name{pmfcont}
\alias{pmfcont}
\alias{pmfcont.default}
\alias{print.pmfcont}
\alias{plot.pmfcont}
\title{pmfcont}
\usage{
pmfcont(x, ...)

\method{pmfcont}{default}(dat1, dir = "./", rms = NULL, prefix,
  constrain = F, cn = NULL, scale = NULL,
  formatdate = "\%m/\%d/\%Y")

\method{print}{pmfcont}(x)

\method{plot}{pmfcont}(x, class = "scatter", type = F, scale = NULL,
  size1 = 12, size2 = 1, ncol1 = 3)
}
\arguments{
\item{dat1}{full path to original constituent dataset}

\item{dir}{path to PMF result}

\item{rms}{character vector of constituents removed from PMF run}

\item{prefix}{prefix of PMF result}

\item{constrain}{was constrained PMF run? (default = F)}

\item{cn}{optional column names for sources}

\item{scale}{character of scaling variable (e.g. PM)}

\item{formatdate}{date format if first column is date}
}
\description{
Reading in and formatting contributions data
}
\details{
\code{pmfcont} Read in contribution data

These are functions to read in and format PMF output contributions
}
\examples{
# Save nycdat.csv and PMF output to working directory
cont <- pmfcont("nycdat.csv", prefix = "pmfdatex", formatdate = "\%m/\%d/\%y")
print(cont)
plot(cont)
plot(cont, class = "hist")
summary(cont)
}
