% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmfprof.R
\name{pmfprof}
\alias{pmfprof}
\alias{pmfprof.default}
\alias{print.pmfprof}
\alias{plot.pmfprof}
\title{pmfprof}
\usage{
pmfprof(x, ...)

\method{pmfprof}{default}(dat, dir = "./", rms = NULL, prefix,
  constrain = F, cn = NULL, scale = F)

\method{print}{pmfprof}(x)

\method{plot}{pmfprof}(x, scale = T, type = F, size = 12)
}
\arguments{
\item{dat}{full path to original constituent dataset}

\item{dir}{path to PMF result}

\item{rms}{character vector of constituents removed from PMF run}

\item{prefix}{prefix of PMF result}

\item{constrain}{was constrained PMF run? (default = F)}

\item{cn}{optional column names for sources}

\item{scale}{whether to scale profiles (default = F)}
}
\description{
Reading in and formatting profile data
}
\details{
\code{pmfprof} Read in profile data

These are functions to read in and format PMF output profiles
}
\examples{
# Save nycdat.csv and PMF output to working directory
prof <- pmfprof("nycdat.csv", prefix = "pmfdatex")
print(prof)
plot(prof)
summary(prof)
}
