% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_base_error_bootstraps.R
\name{read_pmf_base_error_bootstraps}
\alias{read_pmf_base_error_bootstraps}
\title{Function to read EPA PMF base model bootstrapped errors and statistics.}
\usage{
read_pmf_base_error_bootstraps(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
A named list containing tibbles.
}
\description{
Function to read EPA PMF base model bootstrapped errors and statistics.
}
\author{
Stuart K. Grange
}
