% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_base_error_estimation.R
\name{read_pmf_base_error_estimations}
\alias{read_pmf_base_error_estimations}
\title{Function to read PMF base error estimation summarries exported from the EPA
PMF tool.}
\usage{
read_pmf_base_error_estimations(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
Tibble.
}
\description{
Function to read PMF base error estimation summarries exported from the EPA
PMF tool.
}
\author{
Stuart K. Grange
}
