% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_configuration.R
\name{read_pmf_configuration}
\alias{read_pmf_configuration}
\title{Function to read EPA PMF configuration file.}
\usage{
read_pmf_configuration(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
Tibble.
}
\description{
Function to read EPA PMF configuration file.
}
\author{
Stuart K. Grange
}
