% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_constrained_diagnostics.R
\name{read_pmf_constrained_diagnostics}
\alias{read_pmf_constrained_diagnostics}
\title{Function to read constrained PMF diagnostics exported from the EPA PMF tool.}
\usage{
read_pmf_constrained_diagnostics(file, tz = "UTC")
}
\arguments{
\item{file}{File to read.}

\item{tz}{Time zone the dates are stored in.}
}
\value{
Named list containing tibbles.
}
\description{
Function to read constrained PMF diagnostics exported from the EPA PMF tool.
}
\author{
Stuart K. Grange
}
