% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_contributions.R
\name{read_pmf_contributions}
\alias{read_pmf_contributions}
\title{Function to read PMF source contributions exported from the EPA PMF tool.}
\usage{
read_pmf_contributions(file, tz = "UTC")
}
\arguments{
\item{file}{File to read.}

\item{tz}{Time-zone dates are stored in.}
}
\value{
Tibble.
}
\description{
Function to read PMF source contributions exported from the EPA PMF tool.
}
\author{
Stuart K. Grange
}
