% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_f_peak_error_estimation.R
\name{read_pmf_f_peak_error_estimation}
\alias{read_pmf_f_peak_error_estimation}
\title{Function to read PMF F-peak error estimations exported from the EPA PMF tool.}
\usage{
read_pmf_f_peak_error_estimation(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
Named list containing tibbles.
}
\description{
Function to read PMF F-peak error estimations exported from the EPA PMF tool.
}
\author{
Stuart K. Grange
}
