% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_factor_profiles.R
\name{read_pmf_factor_profiles}
\alias{read_pmf_factor_profiles}
\title{Function to read PMF factor profiles exported from the EPA PMF tool.}
\usage{
read_pmf_factor_profiles(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
Tibble.
}
\description{
Function to read PMF factor profiles exported from the EPA PMF tool.
}
\author{
Stuart K. Grange
}
