% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_inputs.R
\name{read_pmf_inputs}
\alias{read_pmf_inputs}
\title{Function to read EPA PMF inputs.}
\usage{
read_pmf_inputs(file_concentrations, file_errors, tz = "UTC")
}
\arguments{
\item{file_concentrations}{File name of concentration file.}

\item{file_errors}{File name of error file. This is optional.}

\item{tz}{Time zone of dates.}
}
\value{
Tibble.
}
\description{
Function to read EPA PMF inputs.
}
\author{
Stuart K. Grange
}
