% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_run_comparison.R
\name{read_pmf_run_comparison}
\alias{read_pmf_run_comparison}
\title{Function to read PMF run comparisions exported from the EPA PMF tool.}
\usage{
read_pmf_run_comparison(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
Tibble.
}
\description{
Function to read PMF run comparisions exported from the EPA PMF tool.
}
\author{
Stuart K. Grange
}
