% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_source_contributions.R
\name{read_pmf_source_contributions}
\alias{read_pmf_source_contributions}
\title{Function to read PMF source contributions statistics exported from the EPA 
PMF tool.}
\usage{
read_pmf_source_contributions(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
Tibble.
}
\description{
Function to read PMF source contributions statistics exported from the EPA 
PMF tool.
}
\author{
Stuart K. Grange
}
