% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_missing_and_negative_values.R
\name{replace_missing_and_negative_values}
\alias{replace_missing_and_negative_values}
\title{Function to replace missing and negative values in concentration and error
tables in preparation for use with the EPA-PMF tool.}
\usage{
replace_missing_and_negative_values(
  df,
  df_errors,
  zero_replace = 1e-06,
  na_replace = 10
)
}
\arguments{
\item{df}{Data frame containing concentration data.}

\item{df_errors}{Data frame containing errror or uncertainty data.}

\item{zero_replace}{Replacement value for values which are <= 0.}

\item{na_replace}{Replacement value for missing values (NA).}
}
\value{
Named list containing two tibbles.
}
\description{
Function to replace missing and negative values in concentration and error
tables in preparation for use with the EPA-PMF tool.
}
\author{
Stuart K. Grange.
}
