% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_base_error_bootstraps.R
\name{tidy_bootstrap_mapping_table}
\alias{tidy_bootstrap_mapping_table}
\title{Function to tidy EPA PMF's bootstrap mapping tables.}
\usage{
tidy_bootstrap_mapping_table(df, pairs_only = FALSE)
}
\arguments{
\item{df}{Bootstrap mapping tibble from 
\code{\link{read_pmf_base_error_bootstraps}} or 
\code{\link{read_pmf_constrained_error_bootstraps}}.}

\item{pairs_only}{Should the return be filtered only to vaiable pairs?}
}
\value{
Tibble.
}
\description{
Function to tidy EPA PMF's bootstrap mapping tables.
}
\author{
Stuart K. Grange
}
