% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_contributions.R
\name{tidy_pmf_contributions}
\alias{tidy_pmf_contributions}
\title{Function to reshape PMF contributions into tidy data.}
\usage{
tidy_pmf_contributions(df)
}
\arguments{
\item{df}{Tibble/data frame from \code{\link{read_pmf_contributions}}.}
}
\value{
Tibble.
}
\description{
Function to reshape PMF contributions into tidy data.
}
\author{
Stuart K. Grange
}
