% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_residuals.R
\name{tidy_pmf_residuals}
\alias{tidy_pmf_residuals}
\title{Function to reshape PMF residuals into tidy data.}
\usage{
tidy_pmf_residuals(df)
}
\arguments{
\item{df}{Tibble/data frame from \code{\link{read_pmf_residuals}}.}
}
\value{
Tibble.
}
\description{
Function to reshape PMF residuals into tidy data.
}
\author{
Stuart K. Grange
}
