% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pmf_mass_factor_contributions.R
\name{extract_pmf_mass_factor_contributions}
\alias{extract_pmf_mass_factor_contributions}
\title{Function to extract PMF factor contributions from the return of
\code{\link{read_pmf_factor_profiles}}.}
\usage{
extract_pmf_mass_factor_contributions(df, value_mass = NA)
}
\arguments{
\item{df}{Tibble from \code{\link{read_pmf_factor_profiles}}.}

\item{value_mass}{An optional value to use for fully accounted for mass in 
the calculations.}
}
\value{
Tibble.
}
\description{
Function to extract PMF factor contributions from the return of
\code{\link{read_pmf_factor_profiles}}.
}
\seealso{
\code{\link{read_pmf_factor_profiles}}, 
\code{\link{plot_pmf_mass_factor_contributions}}
}
\author{
Stuart K. Grange
}
