% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pmf_mass_factor_contributions.R
\name{extract_pmf_mass_factor_contributions}
\alias{extract_pmf_mass_factor_contributions}
\title{Function to extract PMF factor contributions from the return of
\code{\link{read_pmf_source_contributions}}.}
\usage{
extract_pmf_mass_factor_contributions(df, model_run, species = "mass")
}
\arguments{
\item{df}{Tibble from \code{\link{read_pmf_source_contributions}}.}

\item{model_run}{Model run extract contributions from.}

\item{species}{Which species to extract contributions for. Defaults to mass.}
}
\value{
Tibble.
}
\description{
Function to extract PMF factor contributions from the return of
\code{\link{read_pmf_source_contributions}}.
}
\seealso{
\code{\link{read_pmf_source_contributions}}, 
\code{\link{plot_pmf_mass_factor_contributions}}
}
\author{
Stuart K. Grange
}
