% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmf_factor_profiles.R
\name{tidy_pmf_profiles}
\alias{tidy_pmf_profiles}
\title{Function to reshape PMF factor profiles into tidy data.}
\usage{
tidy_pmf_profiles(df)
}
\arguments{
\item{df}{Tibble/data frame from \code{\link{read_pmf_factor_profiles}}.}
}
\value{
Tibble.
}
\description{
Function to reshape PMF factor profiles into tidy data.
}
\author{
Stuart K. Grange
}
